use glob async function
this commit fixes the issue that we saw in the previous commit, where we
no longer use a default export from `glob` (as they no longer exist) and
kinda naively convert our `asyncGlob` function to be a dumb wrapper
around the async glob fn. we can clean this up later - right now, we're
working through type checker/build errors.

when we run the build again, we hit a new error in rollup:
```
npm run build

> @stencil/core@4.0.3 build
> npm run tsc.scripts && npm run tsc.prod && npm run rollup.prod.ci

> @stencil/core@4.0.3 tsc.scripts
> tsc --build --force scripts/tsconfig.json

> @stencil/core@4.0.3 tsc.prod
> tsc

> @stencil/core@4.0.3 rollup.prod.ci
> rollup --config --config-prod --config-ci

[!] TypeError: Cannot read properties of undefined (reading 'sync')
TypeError: Cannot read properties of undefined (reading 'sync')
    at copyPolyfills (/Users/ryan/workspaces/stencil/scripts/build/bundles/internal-platform-client.js:103:45)
    at internalClient (/Users/ryan/workspaces/stencil/scripts/build/bundles/internal-platform-client.js:21:11)
    at internal (/Users/ryan/workspaces/stencil/scripts/build/bundles/internal.js:31:34)
    at async Promise.all (index 3)
    at createBuild (/Users/ryan/workspaces/stencil/scripts/build/build.js:57:21)
    at getConfigList (/Users/ryan/workspaces/stencil/node_modules/rollup/dist/shared/loadConfigFile.js:564:20)
    at Object.loadAndParseConfigFile (/Users/ryan/workspaces/stencil/node_modules/rollup/dist/shared/loadConfigFile.js:476:21)
    at getConfigs (/Users/ryan/workspaces/stencil/node_modules/rollup/dist/bin/rollup:1685:39)
    at runRollup (/Users/ryan/workspaces/stencil/node_modules/rollup/dist/bin/rollup:1659:43)
```

we'll look into that in the next commit